api/requirements.txt
fastapi
uvicorn[standard]
cassandra-driver
python-dotenv
redis